+++ src/main/kotlin/klib/net/socket/KSocket.kt
@@ -89,7 +89,8 @@ class KSocket {
                     }
                 }
             },
-            timeout, timeout
+            timeout,
+            timeout
         )
     }