+++ CHANGELOG.md
@@ -13,12 +13,12 @@ and this project adheres to [Semantic Versioning](semver.md).
 - `mod97` extension for String
 ### Changed
 - Upgraded Kotlin from 1.3.71 to 1.4.0
+- FFDB v1 files are read-only. (FFDB is still Experimental so this change is acceptable)
 ### Deprecated
 - Queue (`klib.queue.Queue`)
 - Function (`klib.queue.Function`)
 - Function -> Any (`klib.queue.Function_Any``)
 - Word (`klib.word.Word`)
-- Writing of FFDB v1 files. v1 will be made read-only in the future.
 ### Removed
 ### Fixed
 ### Security